+++ CHANGELOG.md
@@ -8,6 +8,13 @@ to a section with the version name.
 
 ## Unreleased Changes
 
+* Ignore permissions when inserting ZATCA integration log
+* Fix item total amount in sales invoice additional fields for return invoices
+* Abort submission for sales invoice additional field document if the status is resend.
+* Update ZATCA integration log with Zatca status.
+* Autoname method for ZATCA integration log.
+* Add Last Attempt field in additional field doctype.
+
 ## 0.6.0
 
 * Add support for precomputed invoices from POS devices
@@ -21,13 +28,12 @@ to a section with the version name.
   * Sort additional fields by creation (oldest first)
   * Run in batches (of 100 by default)
   * Add more logging
-* Ignore permissions when inserting ZATCA integration log
-* Fix item total amount in sales invoice additional fields for return invoices
+
 
 ## 0.4.0
 
 * Ignore permissions when creating sales invoice additional fields
-* Skip additional fiels for invoices issued before 2024-03-01
+* Skip additional fields for invoices issued before 2024-03-01
 * Add a flag to control ZATCA XML validation and make it disabled by default
 * Switch signed invoice XML from an attachment to a field for performance reasons